Update logging in dataset manager and adjust target type in validation step for consistency

+++ src/hmc/train/local_classifier/baseline/valid.py
@@ -49,7 +49,7 @@ def valid_step(args):
             for index in args.active_levels:
                 if args.level_active[index]:
                     output = outputs[str(index)]
-                    target = targets[index].float()
+                    target = targets[index].double()
                     loss = args.criterions[index](output, target)
                     local_val_losses[index] += loss
